requirements.txt
django~=3.1.7
django-ckeditor~=6.0.0
django-richtextfield~=1.6
djangorestframework~=3.12.2